hub/ph-api/consumers/thumbnails/npz_thumbnail.py
from loguru import logger
from typing import Dict, Any, Optional
from sqlalchemy import select

from models import File, Project
from services.storage import get_storage_backend
from db import SessionLocal

# In a real implementation, you would use matplotlib to generate plots from numpy data
# For this stub, we'll just return a placeholder


async def generate_npz_thumbnail(file: File) -> Optional[Dict[str, Any]]:
    """
    Generate an image thumbnail for an NPZ (NumPy compressed) file.

    Args:
        file: The File object representing the NPZ file

    Returns:
        A dictionary containing the thumbnail data or None if generation fails
    """
    try:
        # Get the project if file is associated with one
        db = SessionLocal()
        try:
            project = None
            if file.project_id:
                project_stmt = select(Project).where(Project.id == file.project_id)
                project = db.execute(project_stmt).scalar_one_or_none()
        finally:
            db.close()

        # Get the appropriate storage backend based on project settings and file path
        storage_backend = get_storage_backend(project, file.file_path)

        # Check if the file exists by trying to get it from storage
        try:
            # Just check if we can get the file, don't actually read it yet
            await storage_backend.get_file(file.file_path)
        except Exception as e:
            logger.error(
                f"NPZ file not found or inaccessible: {file.file_path}, error: {str(e)}"
            )
            return None

        # In a real implementation, you would:
        # 1. Get the file from storage and save it to a temporary file
        # 2. Load the NPZ file using numpy from the temporary file
        # 3. Extract the data arrays
        # 4. Generate a plot or visualization using matplotlib
        # 5. Save the plot to a buffer
        # 6. Convert the buffer to base64 for storage
        # 7. Clean up the temporary file

        # For this stub, we'll just return a placeholder
        # In a real implementation, replace this with actual NPZ data visualization
        placeholder = "NPZ_PLOT_PLACEHOLDER_BASE64_ENCODED"

        # Return the thumbnail data
        return {
            "thumbnail": placeholder,
            "type": "image",
            "format": "base64",
            "width": 256,  # Placeholder values
            "height": 192,
            "content_type": "image/png",
        }

    except Exception as e:
        logger.error(f"Error generating NPZ thumbnail: {str(e)}")
        return None


# Example of how this would be implemented with actual libraries:
"""
import matplotlib.pyplot as plt
import io

async def generate_npz_thumbnail_real(file: File) -> Optional[Dict[str, Any]]:
    try:
        # Load the NPZ file
        data = np.load(file.file_path)
        
        # Get the array names
        array_names = list(data.keys())
        
        if not array_names:
            logger.error(f"No arrays found in NPZ file: {file.file_path}")
            return None
            
        # Get the first array
        array_name = array_names[0]
        array_data = data[array_name]
        
        # Create a figure
        plt.figure(figsize=(6, 4), dpi=100)
        
        # Plot the data based on its shape
        if len(array_data.shape) == 1:
            # 1D array - line plot
            plt.plot(array_data)
            plt.title(f"1D Array: {array_name}")
            plt.xlabel("Index")
            plt.ylabel("Value")
        elif len(array_data.shape) == 2:
            # 2D array - heatmap
            plt.imshow(array_data, cmap='viridis')
            plt.colorbar()
            plt.title(f"2D Array: {array_name}")
        else:
            # Higher dimensional array - show first slice as heatmap
            plt.imshow(array_data[0], cmap='viridis')
            plt.colorbar()
            plt.title(f"{len(array_data.shape)}D Array: {array_name} (first slice)")
        
        # Save the plot to a buffer
        buf = io.BytesIO()
        plt.savefig(buf, format='png', bbox_inches='tight')
        plt.close()
        
        # Convert to base64
        buf.seek(0)
        img_str = base64.b64encode(buf.getvalue()).decode()
        
        # Return the thumbnail data
        return {
            "thumbnail": img_str,
            "type": "image",
            "format": "base64",
            "width": 600,  # These would be the actual dimensions of the generated plot
            "height": 400,
            "content_type": "image/png",
            "arrays": array_names
        }
        
    except Exception as e:
        logger.error(f"Error generating NPZ thumbnail: {str(e)}")
        return None
"""
