hub/ph-api/pyproject.toml
[tool.poetry]
name = "fastapi-template"
version = "0.1.0"
description = "A generic FastAPI template with authentication, database, and Kafka integration"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.2"}
uvicorn = "^0.32.0"
pydantic-settings = "^2.6.0"
httpx = "^0.27.2"
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"
sqlalchemy = {extras = ["postgresql-asyncpg"], version = "^2.0.36"}
twilio = "^9.3.4"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
python-multipart = "^0.0.12"
alembic = "^1.13.3"
loguru = "^0.7.2"
schedule = "^1.2.2"
pytz = "^2024.2"
redis = "^5.0.3"
aioapns = "^3.3.1"
kafka-python-ng = "^2.2.3"
anthropic = "^0.44.0"
numpy = "^2.2.2"
scikit-learn = "^1.6.1"
pandas = "^2.2.3"
firebase-admin = "^6.6.0"
google-auth = "^2.28.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.120.0"
pydicom = "^3.0.1"
fuzzywuzzy = "^0.18.0"
torch = "^2.6.0"
torchvision = "^0.21.0"
python-levenshtein = "^0.27.1"
confluent-kafka = "^2.9.0"
boto3 = "^1.34.0"
neurokit2 = "^0.2.10"

[tool.poetry.group.dev.dependencies]
alembic = "^1.13.3"
ruff = "^0.7.0"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
pytest-retry = "^1.6.3"
pre-commit = "^4.0.1"
reorder-python-imports = "^3.14.0"
pytest-asyncio = "^0.25.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
path = ".venv"

[tool.pytest.ini_options]
minversion = "8"
addopts = "-ra -q"
testpaths = [
    "tests",
    "consumers/tests",
    "services",
    "."
]
python_files = ["test_*.py", "*_test.py"]
